+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+## 0.2.3
+
+- Allow `ServiceAccountCredentials` constructors to take an optional
+  `user` argument to specify a user to impersonate.
+
 ## 0.2.2
 
 - Allow `ServiceAccountCredentials.fromJson` to accept a `Map`.
 - Cleaned up `README.md`
 
 ## 0.2.1
-- Added optinoal `force` and `immediate` arguments to `runHybridFlow`.
+- Added optional `force` and `immediate` arguments to `runHybridFlow`.
 
 ## 0.2.0
 - Renamed `forceUserConsent` parameter to `immediate`.